+++ spec/spec_helper.rb
@@ -34,24 +34,5 @@ RSpec.configure do |config|
     $test_infrastructure.convert_all!(true)
   end
 
-  # Disabled this until I figured out why it stopped the coverage
-  # report from being written
-  #
-  # # Enforce 100% test coverage unless we're running in guard
-  # config.after(:suite) do
-  #   example_group = RSpec.describe('Code coverage')
-  #   example = example_group.example('must be 100%'){
-  #     expect( SimpleCov.result.covered_percent ).to eq 100
-  #   }
-  #   example_group.run
-
-  #   passed = example.execution_result.status == :passed
-  #   if passed or ENV['IGNORE_LOW_COVERAGE']
-  #     RSpec.configuration.reporter.example_passed example
-  #   else
-  #     RSpec.configuration.reporter.example_failed example
-  #   end
-  # end
-
   config.order = :random
 end